finished bottom feeder

+++ orderbook.h
@@ -11,7 +11,7 @@ struct Trade {
     Player* buyer;
     Player* seller;
     int suit;
-    int price;
+    float price;
     Trade(Player* b, Player* s, int sType, int p): buyer(b), seller(s), suit(sType), price(p) {}
 };
 
@@ -29,7 +29,7 @@ class orderBook {
         Order* getBestAsk();
         float getBestBidPrice();
         float getBestAskPrice();
-        void executeTrade(int suite, int tradePrice, Player* bidder, Player* asker);
+        void executeTrade(int suite, float tradePrice, Player* bidder, Player* asker);
         void printBook();
         std::vector<Trade> getTrades();
 };
